BinderEventListener should wait for events to arrive
Merely selecting values of the BindEvents dictionary will trigger an
exception if an event arrives after calling the Where() method.

Provide a WaitForEventsForAssembly() method, which will wait until a
timeout is reached (BinderTracingTest defines it to 1000ms), and
selecting not on the BindEvents.Values directly, but on a copy made as
a linked list (this way, the dictionary can be changed without any
issues).